[refactor] api -> actions
