+++ app/(main)/blog/page.tsx
@@ -2,7 +2,7 @@ import styles from "./feed.module.scss";
 import DraftBlogFeed from "@/components/blog/draft-feed";
 import BlogFeed from "@/components/blog/blog-feed";
 import ButtonCreateBlog from "@/components/blog/button-create-blog";
-import { getSession } from "@/api/auth";
+import { getSession } from "@/actions/auth";
 import { Metadata } from "next";
 import SessionRefresher from "@/components/auth/session-refresher";
